Bugfix: profile migration problem, bundle update

+++ db/migrate/20160119101214_add_constraint_to_profile.rb
@@ -1,6 +1,6 @@
 class AddConstraintToProfile < ActiveRecord::Migration
   def change
-    remove_index :profiles, column: :user_id
+    remove_index :profiles, column: :user_id if index_exists?(:profiles, :user_id)
     add_index :profiles, :user_id, unique: true
   end
 end
